+++ packages/frontend/core/src/blocksuite/ai/extensions/index.ts
@@ -1,0 +1,1 @@
+export * from './enable-ai';
