+++ packages/frontend/core/src/blocksuite/ai/index.ts
@@ -1,10 +1,10 @@
 export * from './_common/config';
-export * from './actions/index';
-export * from './ai-spec';
-export { ChatPanel } from './chat-panel/index';
+export * from './actions';
+export { ChatPanel } from './chat-panel';
+export * from './entries';
 export * from './entries/edgeless/actions-config';
-export * from './entries/index';
-export * from './messages/index';
+export * from './extensions';
+export * from './messages';
 export { AIChatBlockPeekViewTemplate } from './peek-view/chat-block-peek-view';
 export * from './provider';
 export * from './utils/edgeless';
