+++ packages/frontend/core/src/blocksuite/block-suite-editor/specs/edgeless.ts
@@ -1,16 +1,12 @@
-import { createAIEdgelessRootBlockSpec } from '@affine/core/blocksuite/ai';
+import { enableAIExtension } from '@affine/core/blocksuite/ai';
 import { FeatureFlagService } from '@affine/core/modules/feature-flag';
 import { builtInTemplates as builtInEdgelessTemplates } from '@affine/templates/edgeless';
 import { builtInTemplates as builtInStickersTemplates } from '@affine/templates/stickers';
 import type { SpecBuilder, TemplateManager } from '@blocksuite/affine/blocks';
-import {
-  EdgelessRootBlockSpec,
-  EdgelessTemplatePanel,
-  SpecProvider,
-} from '@blocksuite/affine/blocks';
+import { EdgelessTemplatePanel, SpecProvider } from '@blocksuite/affine/blocks';
 import { type FrameworkProvider } from '@toeverything/infra';
 
-import { enableAffineExtension, enableAIExtension } from './custom/root-block';
+import { enableAffineExtension } from './custom/root-block';
 
 export function createEdgelessModeSpecs(
   framework: FrameworkProvider
@@ -20,11 +16,7 @@ export function createEdgelessModeSpecs(
   const edgelessSpec = SpecProvider._.getSpec('edgeless');
   enableAffineExtension(framework, edgelessSpec);
   if (enableAI) {
-    enableAIExtension(edgelessSpec);
-    edgelessSpec.replace(
-      EdgelessRootBlockSpec,
-      createAIEdgelessRootBlockSpec(framework)
-    );
+    enableAIExtension(edgelessSpec, framework);
   }
 
   return edgelessSpec;
