+++ packages/frontend/core/src/blocksuite/block-suite-editor/specs/page.ts
@@ -1,13 +1,9 @@
-import { createAIPageRootBlockSpec } from '@affine/core/blocksuite/ai';
+import { enableAIExtension } from '@affine/core/blocksuite/ai';
 import { FeatureFlagService } from '@affine/core/modules/feature-flag';
-import {
-  PageRootBlockSpec,
-  type SpecBuilder,
-  SpecProvider,
-} from '@blocksuite/affine/blocks';
+import { type SpecBuilder, SpecProvider } from '@blocksuite/affine/blocks';
 import { type FrameworkProvider } from '@toeverything/infra';
 
-import { enableAffineExtension, enableAIExtension } from './custom/root-block';
+import { enableAffineExtension } from './custom/root-block';
 
 export function createPageModeSpecs(framework: FrameworkProvider): SpecBuilder {
   const featureFlagService = framework.get(FeatureFlagService);
@@ -16,8 +12,7 @@ export function createPageModeSpecs(framework: FrameworkProvider): SpecBuilder {
   const pageSpec = provider.getSpec('page');
   enableAffineExtension(framework, pageSpec);
   if (enableAI) {
-    enableAIExtension(pageSpec);
-    pageSpec.replace(PageRootBlockSpec, createAIPageRootBlockSpec(framework));
+    enableAIExtension(pageSpec, framework);
   }
   return pageSpec;
 }
